# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"PcmrIndex": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"img": {
					"type": "string",
					"x-nullable": true
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		},
		"Comment": {
			"properties": {
				"body": {
					"type": "string"
				},
				"author": {
					"type": "string"
				},
				"created_at": {
					"type": "string"
				}
			},
			"required": [
				"body",
				"author",
				"created_at"
			],
			"type": "object"
		},
		"PcmrPost": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"img": {
					"type": "string",
					"x-nullable": true
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				},
				"comments": {
					"$ref": "#/definitions/Comment",
					"x-nullable": true
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		},
		"PcGamingIndex": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		},
		"PcGamingPost": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				},
				"comments": {
					"$ref": "#/definitions/Comment",
					"x-nullable": true
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		},
		"BuildPcIndex": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		},
		"BuildPcPost": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				},
				"comments": {
					"$ref": "#/definitions/Comment",
					"x-nullable": true
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		},
		"PcSaleIndex": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		},
		"PcSalePost": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"link": {
					"type": "string"
				},
				"created_at": {
					"type": "string"
				},
				"submissionId": {
					"type": "string"
				},
				"comments": {
					"$ref": "#/definitions/Comment",
					"x-nullable": true
				}
			},
			"required": [
				"_id",
				"title",
				"link",
				"created_at",
				"submissionId"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "pcmr-api",
		"version": "1.0.0",
		"description": "Backend express server for PCMR web client.",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/pcmr": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PcmrIndex"
							}
						}
					}
				},
				"tags": [
					"Pc Master Race"
				],
				"security": [],
				"parameters": []
			}
		},
		"/pcmr/{id}": {
			"get": {
				"operationId": "GetOne",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PcmrPost"
						}
					}
				},
				"tags": [
					"Pc Master Race"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Pc Master Race"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"link": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"link"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Pc Master Race"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/pc-gaming": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PcGamingIndex"
							}
						}
					}
				},
				"tags": [
					"Pc Gaming"
				],
				"security": [],
				"parameters": []
			}
		},
		"/pc-gaming/{id}": {
			"get": {
				"operationId": "GetOne",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PcGamingPost"
						}
					}
				},
				"tags": [
					"Pc Gaming"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Pc Gaming"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"link": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"link"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Pc Gaming"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/build-pc": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/BuildPcIndex"
							}
						}
					}
				},
				"tags": [
					"Build a Pc"
				],
				"security": [],
				"parameters": []
			}
		},
		"/build-pc/{id}": {
			"get": {
				"operationId": "GetOne",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/BuildPcPost"
						}
					}
				},
				"tags": [
					"Build a Pc"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Build a Pc"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"link": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"link"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Build a Pc"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/pc-sales": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PcSaleIndex"
							}
						}
					}
				},
				"tags": [
					"Pc Sales"
				],
				"security": [],
				"parameters": []
			}
		},
		"/pc-sales/{id}": {
			"get": {
				"operationId": "GetOne",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PcSalePost"
						}
					}
				},
				"tags": [
					"Pc Sales"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Pc Sales"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"schema": {
							"properties": {
								"link": {
									"type": "string"
								}
							},
							"title": "UpdateBody",
							"type": "object",
							"required": [
								"link"
							]
						}
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Pc Sales"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}